+++ x-pack/plugin/ml/src/main/java/org/elasticsearch/xpack/ml/autoscaling/NativeMemoryCapacity.java
@@ -11,11 +11,12 @@ import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.elasticsearch.cluster.node.DiscoveryNode;
 import org.elasticsearch.common.unit.ByteSizeValue;
-import org.elasticsearch.xpack.autoscaling.capacity.AutoscalingCapacity;
+import org.elasticsearch.xpack.ml.MachineLearning;
 import org.elasticsearch.xpack.ml.utils.NativeMemoryCalculator;
 
 import java.util.Arrays;
 import java.util.List;
+import java.util.Map;
 import java.util.Objects;
 import java.util.OptionalLong;
 
@@ -107,7 +108,7 @@ public class NativeMemoryCapacity {
      * @return The minimum node size required for ML nodes and the minimum tier size required for the complete ML
      *         tier.
      */
-    public AutoscalingCapacity autoscalingCapacity(
+    public MlMemoryAutoscalingCapacity.Builder autoscalingCapacity(
         int maxMemoryPercent,
         boolean useAuto,
         long mlNativeMemoryForLargestMlNode,
@@ -132,10 +133,7 @@ public class NativeMemoryCapacity {
                     nodeMlNativeMemoryRequirementExcludingOverhead
                 );
             }
-            return new AutoscalingCapacity(
-                new AutoscalingCapacity.AutoscalingResources(null, ByteSizeValue.ZERO, null),
-                new AutoscalingCapacity.AutoscalingResources(null, ByteSizeValue.ZERO, null)
-            );
+            return MlMemoryAutoscalingCapacity.builder(ByteSizeValue.ZERO, ByteSizeValue.ZERO);
         }
 
         if (mlNativeMemoryForLargestMlNode <= NATIVE_EXECUTABLE_CODE_OVERHEAD.getBytes()) {
@@ -223,9 +221,9 @@ public class NativeMemoryCapacity {
         }
         // The assertion above should hold, but the Math.max below catches the case with inconsistent
         // inputs plus any bugs that weren't caught in tests.
-        return new AutoscalingCapacity(
-            new AutoscalingCapacity.AutoscalingResources(null, ByteSizeValue.ofBytes(Math.max(requiredTierSize, requiredNodeSize)), null),
-            new AutoscalingCapacity.AutoscalingResources(null, ByteSizeValue.ofBytes(requiredNodeSize), null)
+        return MlMemoryAutoscalingCapacity.builder(
+            ByteSizeValue.ofBytes(requiredNodeSize),
+            ByteSizeValue.ofBytes(Math.max(requiredTierSize, requiredNodeSize))
         );
     }
 
@@ -290,11 +288,27 @@ public class NativeMemoryCapacity {
             Arrays.stream(mlMemory).max().orElse(0L),
             // We assume that JVM size is universal, at least, the largest JVM indicates the largest node
             machineLearningNodes.stream()
-                .map(MlAutoscalingDeciderService::getNodeJvmSize)
+                .map(NativeMemoryCapacity::getNodeJvmSize)
                 .filter(OptionalLong::isPresent)
                 .map(OptionalLong::getAsLong)
                 .max(Long::compare)
                 .orElse(null)
         );
+    }
+
+    static OptionalLong getNodeJvmSize(DiscoveryNode node) {
+        Map<String, String> nodeAttributes = node.getAttributes();
+        String valueStr = nodeAttributes.get(MachineLearning.MAX_JVM_SIZE_NODE_ATTR);
+        try {
+            return OptionalLong.of(Long.parseLong(valueStr));
+        } catch (NumberFormatException e) {
+            assert e == null : "ml.max_jvm_size should parse because we set it internally: invalid value was " + valueStr;
+            logger.debug(
+                "could not parse stored string value [{}] in node attribute [{}]",
+                valueStr,
+                MachineLearning.MAX_JVM_SIZE_NODE_ATTR
+            );
+        }
+        return OptionalLong.empty();
     }
 }
